feat(reality-web): listing detail renders through resolved layout sections
