+++ frontend/apps/reality-web/src/app/[locale]/listings/[slug]/page.tsx
@@ -16,6 +16,7 @@ import { ListingDetailContent } from '@/components/listings';
 import { buildListingJsonLd } from './jsonLd';
 import { parseListingDetail } from './listingSchema';
 import { buildListingMetadata } from './metadata';
+import { getResolvedLayout } from '@/lib/layout';
 
 function inferApiBaseFromHost(host: string): string | null {
   const bareHost = host.split(':')[0]?.toLowerCase();
@@ -102,7 +103,7 @@ export async function generateMetadata({ params }: PageProps): Promise<Metadata>
 export default async function ListingDetailPage({ params }: PageProps) {
   const { slug } = await params;
   const host = await resolveHost();
-  const listing = await getListing(slug, host);
+  const [listing, layout] = await Promise.all([getListing(slug, host), getResolvedLayout(host)]);
 
   // A missing / malformed listing must emit a real HTTP 404, not a 200 with
   // "not found" markup: this is a public, SEO-indexed portal, and a soft-404
@@ -118,5 +119,5 @@ export default async function ListingDetailPage({ params }: PageProps) {
   // can dereference required fields directly.
   const jsonLd = buildListingJsonLd(listing);
 
-  return <ListingDetailContent listing={listing} jsonLd={jsonLd} />;
+  return <ListingDetailContent listing={listing} jsonLd={jsonLd} layout={layout} />;
 }
