+++ frontend/apps/reality-web/src/components/listings/sections/registry.tsx
@@ -1,5 +1,13 @@
 import type { ListingDetail } from '@ppt/reality-api-client';
 import type { ComponentType } from 'react';
+import { AdditionalInfo } from './AdditionalInfo';
+import { AgentContact } from './AgentContact';
+import { Description } from './Description';
+import { Features } from './Features';
+import { Gallery } from './Gallery';
+import { KeyDetails } from './KeyDetails';
+import { ListingHeader } from './ListingHeader';
+import { Resources } from './Resources';
 
 export interface ListingSectionProps {
   listing: ListingDetail;
@@ -15,8 +23,13 @@ export interface ListingSectionDef {
 
 export type ListingRegistry = Record<string, ListingSectionDef>;
 
-/**
- * Registry of listing-detail section components.
- * Intentionally empty — Task 6 fills in the actual section implementations.
- */
-export const listingRegistry: ListingRegistry = {};
+export const listingRegistry: ListingRegistry = {
+  'gallery.v1': { component: Gallery, required: true, supportedModes: [] },
+  'listing-header.v1': { component: ListingHeader, required: true, supportedModes: [] },
+  'key-details.v1': { component: KeyDetails, required: false, supportedModes: [] },
+  'description.v1': { component: Description, required: false, supportedModes: [] },
+  'features.v1': { component: Features, required: false, supportedModes: [] },
+  'additional-info.v1': { component: AdditionalInfo, required: false, supportedModes: [] },
+  'resources.v1': { component: Resources, required: false, supportedModes: [] },
+  'agent-contact.v1': { component: AgentContact, required: false, supportedModes: [] },
+};
